+++ src/imio/smartweb/common/tests/test_cropping.py
@@ -34,4 +34,4 @@ class TestCropping(unittest.TestCase):
         cropping_view = getMultiAdapter(
             (self.folder, self.request), name="croppingeditor"
         )
-        self.assertEqual(len(list(cropping_view._scales("image"))), 13)
+        self.assertEqual(len(list(cropping_view._scales("image"))), 11)