+++ src/imio/smartweb/common/utils.py
@@ -1,10 +1,16 @@
 # -*- coding: utf-8 -*-
 
 from imio.smartweb.common.config import TRANSLATED_VOCABULARIES
+from imio.smartweb.common.interfaces import ICropping
+from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone import api
+from plone.dexterity.utils import iterSchemata
 from plone.formwidget.geolocation.geolocation import Geolocation
+from plone.namedfile.field import NamedBlobImage
+from plone.namedfile.interfaces import IAvailableSizes
 from zope.component import getUtility
 from zope.i18n import translate
+from zope.schema import getFields
 from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 
@@ -81,3 +87,63 @@ def rich_description(description):
     # <br/>
     description = "<br/>".join(description.split("\r\n"))
     return description
+
+
+def get_uncroppable_scales_infos(image, available_sizes, scales):
+    img_height = image._height
+    img_width = image._width
+    result = {}
+    uncroppable_scales = set()
+    min_height = min_width = 0
+    for scale in scales:
+        width, height = available_sizes.get(scale)
+        if 65536 > width > img_width or 65536 > height > img_height:
+            uncroppable_scales.add(scale)
+        if 65536 > width > min_width:
+            min_width = width
+        if 65536 > height > min_height:
+            min_height = height
+    if uncroppable_scales:
+        result = {
+            "scales": sorted(list(uncroppable_scales)),
+            "min_height": str(min_height),
+            "min_width": str(min_width),
+            "height": str(img_height),
+            "width": str(img_width),
+        }
+    return result
+
+
+def show_warning_for_scales(obj, request):
+    if not api.user.has_permission("Modify portal content", obj=obj):
+        return
+    available_sizes = getUtility(IAvailableSizes)() or {}
+    cropping_scales_adapter = ICropping(obj, alternate=None)
+    for schema in iterSchemata(obj):
+        for name, field in getFields(schema).items():
+            if type(field) is not NamedBlobImage or not getattr(obj, name, None):
+                continue
+            field_scales = cropping_scales_adapter.get_scales(name, request)
+            uncroppable_infos = get_uncroppable_scales_infos(
+                getattr(obj, name), available_sizes, field_scales
+            )
+            if not uncroppable_infos:
+                continue
+            api.portal.show_message(
+                _(
+                    'The image uploaded in the "${field_title}" field may be '
+                    "degraded because it does not meet the required minimum dimensions of "
+                    "${min_width}px width by ${min_height}px height "
+                    "(uploaded image size: ${width}px width by ${height}px height). "
+                    "You can see the detail via the Cropping menu.",
+                    mapping={
+                        "field_title": field.title,
+                        "min_height": uncroppable_infos["min_height"],
+                        "min_width": uncroppable_infos["min_width"],
+                        "height": uncroppable_infos["height"],
+                        "width": uncroppable_infos["width"],
+                    },
+                ),
+                request=request,
+                type="warning",
+            )